+++ praw/decorators.py
@@ -27,6 +27,7 @@ import six
 import sys
 from functools import wraps
 from requests.compat import urljoin
+from requests.exceptions import HTTPError
 from praw import errors
 from warnings import simplefilter, warn
 
@@ -158,7 +159,18 @@ def raise_api_exceptions(function):
     """
     @wraps(function)
     def wrapped(reddit_session, *args, **kwargs):
-        return_value = function(reddit_session, *args, **kwargs)
+        try:
+            return_value = function(reddit_session, *args, **kwargs)
+        except HTTPError as exc:
+            if exc.response.status_code != 400:
+                raise  # Unhandled HTTPErrors
+            try:  # Attempt to convert v1 errors into older format (for now)
+                data = exc.response.json()
+                assert len(data) == 2
+                return_value = {'errors': [(data['reason'],
+                                            data['explanation'], '')]}
+            except Exception:
+                raise exc
         if isinstance(return_value, dict):
             if return_value.get('error') == 304:  # Not modified exception
                 raise errors.NotModified(return_value)